+++ tests/app/models/__init__.py
@@ -1,2 +1,4 @@
+from .dummy import *
 from .base import *
+from .custom import *
 from .site import *